feat(Home): refactor carousel cells to be more Swifty
